backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
python-multipart==0.0.9
openpyxl==3.1.2
python-dotenv==1.0.1
